Nueva hab, nuevo hot

+++ AppReservas/AppReservas/Models/Habitacion.cs
@@ -10,7 +10,7 @@ namespace AppReservas.Models
         public int HAB_CODIGO { get; set; }
         public int HAB_CANT_HUESP { get; set; }
         public string HAB_TIPO { get; set; }
-        public int HAB_PRECIO { get; set; }
+        public decimal HAB_PRECIO { get; set; }
         public string HAB_ESTADO { get; set; }
         public int HOT_CODIGO { get; set; }
     }
